+++ packages/content-editor-plugin/src/components/Dialog.css.ts
@@ -1,0 +1,7 @@
+import { style } from '@vanilla-extract/css';
+
+export default {
+  root: style({
+    minWidth: 500
+  })
+};