+++ packages/content-editor-plugin/src/components/Toolbar/InsertHorizontalRule.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import { ToolbarButton } from '@salt-ds/lab';
 import { Icon } from '@jpmorganchase/mosaic-components';
 import { useLexicalComposerContext } from '@lexical/react/LexicalComposerContext';
 import { INSERT_HORIZONTAL_RULE_COMMAND } from '@lexical/react/LexicalHorizontalRuleNode';
+import { ToolbarButton } from './ToolbarButton';
 
 export const InsertHorizontalRule = () => {
   const [editor] = useLexicalComposerContext();
@@ -11,7 +11,7 @@ export const InsertHorizontalRule = () => {
     editor.dispatchCommand(INSERT_HORIZONTAL_RULE_COMMAND, undefined);
   };
   return (
-    <ToolbarButton onClick={handleClick}>
+    <ToolbarButton onClick={handleClick} label="Insert Divider">
       <Icon name="minimize" />
     </ToolbarButton>
   );