+++ packages/content-editor-plugin/src/components/Toolbar/TextFormatTooltray.tsx
@@ -15,8 +15,8 @@ import { $isHeadingNode } from '@lexical/rich-text';
 import { $isCodeNode } from '@lexical/code';
 import { Icon } from '@jpmorganchase/mosaic-components';
 
-import { Tooltray } from '@salt-ds/lab';
-import { TextFormatToolbarButton } from './TextFormatToolbarButton';
+import { BaseTooltray as Tooltray } from '../BaseTooltray/BaseTooltray';
+import { ToolbarButton } from './ToolbarButton';
 import styles from './TextFormatTooltray.css';
 import { InsertBlockDropdown } from './InsertBlockDropdown';
 
@@ -88,31 +88,26 @@ export function TextFormatTooltray({ floating = false }) {
   );
 
   return (
-    <div className={styles.root}>
-      {!floating ? <span /> : null}
-      <Tooltray data-collapsible="instant">
-        {!floating && <InsertBlockDropdown editor={activeEditor} type={blockType} />}
-        <TextFormatToolbarButton active={isBold} onClick={() => handleFormat('bold')} label="Bold">
+    <>
+      {!floating && (
+        <Tooltray>
+          <InsertBlockDropdown editor={activeEditor} type={blockType} />
+        </Tooltray>
+      )}
+      <Tooltray>
+        <ToolbarButton active={isBold} onClick={() => handleFormat('bold')} label="Bold">
           <span className={classnames(styles.icon, styles.bold)}>B</span>
-        </TextFormatToolbarButton>
-        <TextFormatToolbarButton
-          active={isItalic}
-          onClick={() => handleFormat('italic')}
-          label="Italic"
-        >
+        </ToolbarButton>
+        <ToolbarButton active={isItalic} onClick={() => handleFormat('italic')} label="Italic">
           <span className={classnames(styles.icon, styles.italic)}>I</span>
-        </TextFormatToolbarButton>
-        <TextFormatToolbarButton
-          active={isCode}
-          onClick={() => handleFormat('code')}
-          label="Inline Code"
-        >
+        </ToolbarButton>
+        <ToolbarButton active={isCode} onClick={() => handleFormat('code')} label="Inline Code">
           <span className={classnames(styles.code)}>
             <Icon name="chevronLeft" />
             <Icon name="chevronRight" />
           </span>
-        </TextFormatToolbarButton>
+        </ToolbarButton>
       </Tooltray>
-    </div>
+    </>
   );
 }