+++ packages/content-editor-plugin/src/plugins/FloatingToolbarPlugin.css.ts
@@ -1,0 +1,14 @@
+import { vars } from '@jpmorganchase/mosaic-theme';
+import { globalStyle, style } from '@vanilla-extract/css';
+
+const toolbar = style({
+  minWidth: '100px'
+});
+
+export default {
+  toolbar
+};
+
+globalStyle(`${toolbar} > *:not(:first-child)`, {
+  marginLeft: vars.space.horizontal.none
+});