+++ packages/site-middleware/src/__tests__/withSharedConfig.test.ts
@@ -1,6 +1,6 @@
 import { GetObjectCommand, HeadObjectCommand, S3Client } from '@aws-sdk/client-s3';
 import { AwsStub, mockClient } from 'aws-sdk-client-mock';
-import { sdkStreamMixin } from '@aws-sdk/util-stream-node';
+import { sdkStreamMixin } from '@aws-sdk/util-stream';
 import { Readable } from 'stream';
 import { default as fetchMock, disableFetchMocks, enableFetchMocks } from 'jest-fetch-mock';
 const mockFs = require('mock-fs');