+++ packages/content-editor-plugin/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @jpmorganchase/mosaic-content-editor-plugin
 
+## 0.1.0-beta.38
+
+### Patch Changes
+
+- dc993df: Styling and API changes for In-Browser Content Editor (IBCE)
+- Updated dependencies [18f7a0b]
+  - @jpmorganchase/mosaic-theme@0.1.0-beta.38
+  - @jpmorganchase/mosaic-components@0.1.0-beta.38
+
+## 0.1.0-beta.37
+
+### Patch Changes
+
+- @jpmorganchase/mosaic-components@0.1.0-beta.37
+- @jpmorganchase/mosaic-theme@0.1.0-beta.37
+
 ## 0.1.0-beta.36
 
 ### Patch Changes