+++ packages/core/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @jpmorganchase/mosaic-core
 
+## 0.1.0-beta.38
+
+### Patch Changes
+
+- dc993df: Styling and API changes for In-Browser Content Editor (IBCE)
+  - @jpmorganchase/mosaic-plugins@0.1.0-beta.38
+  - @jpmorganchase/mosaic-schemas@0.1.0-beta.38
+  - @jpmorganchase/mosaic-serialisers@0.1.0-beta.38
+  - @jpmorganchase/mosaic-types@0.1.0-beta.38
+
+## 0.1.0-beta.37
+
+### Patch Changes
+
+- Updated dependencies [feb5b9e]
+  - @jpmorganchase/mosaic-plugins@0.1.0-beta.37
+  - @jpmorganchase/mosaic-schemas@0.1.0-beta.37
+  - @jpmorganchase/mosaic-serialisers@0.1.0-beta.37
+  - @jpmorganchase/mosaic-types@0.1.0-beta.37
+
 ## 0.1.0-beta.36
 
 ### Patch Changes