+++ packages/open-api-component/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @jpmorganchase/mosaic-open-api-component
 
+## 0.1.0-beta.38
+
+### Patch Changes
+
+- Updated dependencies [18f7a0b]
+  - @jpmorganchase/mosaic-theme@0.1.0-beta.38
+  - @jpmorganchase/mosaic-components@0.1.0-beta.38
+
+## 0.1.0-beta.37
+
+### Patch Changes
+
+- @jpmorganchase/mosaic-components@0.1.0-beta.37
+- @jpmorganchase/mosaic-theme@0.1.0-beta.37
+
 ## 0.1.0-beta.36
 
 ### Patch Changes