+++ packages/source-http/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @jpmorganchase/mosaic-source-http
 
+## 0.1.0-beta.38
+
+### Patch Changes
+
+- @jpmorganchase/mosaic-from-http-request@0.1.0-beta.38
+- @jpmorganchase/mosaic-schemas@0.1.0-beta.38
+- @jpmorganchase/mosaic-types@0.1.0-beta.38
+
+## 0.1.0-beta.37
+
+### Patch Changes
+
+- @jpmorganchase/mosaic-from-http-request@0.1.0-beta.37
+- @jpmorganchase/mosaic-schemas@0.1.0-beta.37
+- @jpmorganchase/mosaic-types@0.1.0-beta.37
+
 ## 0.1.0-beta.36
 
 ### Patch Changes