+++ packages/types/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @jpmorganchase/mosaic-types
 
+## 0.1.0-beta.38
+
+## 0.1.0-beta.37
+
 ## 0.1.0-beta.36
 
 ## 0.1.0-beta.35