+++ packages/workflows/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @jpmorganchase/mosaic-workflows
 
+## 0.1.0-beta.38
+
+### Patch Changes
+
+- dc993df: Styling and API changes for In-Browser Content Editor (IBCE)
+  - @jpmorganchase/mosaic-serialisers@0.1.0-beta.38
+  - @jpmorganchase/mosaic-source-git-repo@0.1.0-beta.38
+  - @jpmorganchase/mosaic-types@0.1.0-beta.38
+
+## 0.1.0-beta.37
+
+### Patch Changes
+
+- @jpmorganchase/mosaic-serialisers@0.1.0-beta.37
+- @jpmorganchase/mosaic-source-git-repo@0.1.0-beta.37
+- @jpmorganchase/mosaic-types@0.1.0-beta.37
+
 ## 0.1.0-beta.36
 
 ### Patch Changes